+++ ext/JEMExt/JEMExt.jl
@@ -1,0 +1,10 @@
+module JEMExt
+
+using CounterfactualExplanations
+using CounterfactualExplanations.Models: Models
+using JointEnergyModels: JointEnergyModels
+
+include("jem.jl")
+include("pretrained.jl")
+
+end