+++ src/global_utils.jl
@@ -89,7 +89,7 @@ function (encoder::OutputEncoder)(; return_y::Bool=true)
     if !return_y
         return y_levels, likelihood
     else
-        y = Int.(y.refs)
+        y = Int.(encoder.labels.refs)
         if likelihood == :classification_binary
             y = permutedims(y)
             y = y .- 1  # map to [0,1]
@@ -172,4 +172,16 @@ function reset!(flux_training_params::FluxModelParams)
         )
     end
     return flux_training_params
+end
+
+"""
+    polynomial_decay(a::Real, b::Real, decay::Real, t::Int)
+
+Computes the polynomial decay function as in Welling et al. (2011): https://www.stats.ox.ac.uk/~teh/research/compstats/WelTeh2011a.pdf.
+"""
+function polynomial_decay(a::Real, b::Real, decay::Real, t::Int)
+    if decay <= 0.5 || decay > 1.0
+        @warn "Decay rate should be in the range (0.5, 1.0]. See Welling et al. (2011) for more information: https://www.stats.ox.ac.uk/~teh/research/compstats/WelTeh2011a.pdf."
+    end
+    return a * (b + t)^(-decay)
 end