+++ CHANGELOG.md
@@ -10,7 +10,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
-- Added support for an energy constraint as in Altmeyer et al. ([2024](https://scholar.google.com/scholar?cluster=3697701546144846732&hl=en&as_sdt=0,5)). This is the first step towards adding functionality for ECCCo. [387] 
+- Added out-of-the-box support for training joint energy models (JEM). [#454]
+- Added new evaluation metric to measure faithfulness of counterfactual explanations as in Altmeyer et al. ([2024](https://scholar.google.com/scholar?cluster=3697701546144846732&hl=en&as_sdt=0,5)). [#454]
+- A tutorial in the documentation ("Explanation" section) explaining the faithfulness metric in detail. [#454]
+- Added support for an energy constraint as in Altmeyer et al. ([2024](https://scholar.google.com/scholar?cluster=3697701546144846732&hl=en&as_sdt=0,5)). This is the first step towards adding functionality for ECCCo. [#387] 
+  
+### Changed
+
+- The `fitresult` field of `Model` now takes a concrete `Fitresult` type, for which some basic methods have been defined. This mutable struct has a field called `other` that accepts a dictionary `Dict` that can be filled with additional objects. [#454]
+- Regenerated pre-trained model artifacts. [#454]
+- Updated the tutorial on "Handling Data". [#454]
+
+### Removed
+
+- Removed bug in `find_potential_neighbours` method. [#454]
 
 ## Version [1.1.6] - 2024-05-19