+++ docs/src/release-notes.md
@@ -10,10 +10,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 *Note*: We try to adhere to these practices as of version [v1.1.1](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/releases/tag/v1.1.1).
 
-## Version [1.1.6] - 2024-05-16
+## Version [1.2.0] - 2024-05-23
+
+### Added
+
+- Added new evaluation metric to measure unfaithfulness as in Altmeyer et al. ([2024](https://scholar.google.com/scholar?cluster=3697701546144846732&hl=en&as_sdt=0,5)). [#454](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/454)
+- Added support for an energy constraint as in Altmeyer et al. ([2024](https://scholar.google.com/scholar?cluster=3697701546144846732&hl=en&as_sdt=0,5)). This is the first step towards adding functionality for ECCCo. [#387](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/387) 
 
 ### Removed
 
+- Removed bug in `find_potential_neighbours` method. [#454](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/454)
+
+## Version [1.1.6] - 2024-05-19
+
+### Removed
+
+- Removed the call to the `Iris` function in the test suite because of HTTPs issues. [#452](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/452)
 - Removed the `mlj_models_catalogue` because it served no obvious purpose. In the future, we may instead add meta information to the `all_models_catalogue`. [#444](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/444)
 
 ### Added
@@ -23,6 +35,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Changed
 
+- No longer exporting many of the deprecated functions. [#452](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/452)
+- Updated pre-trained model artifacts. [#444](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/444)
 - Some function signatures have been deprecated, e.g. `NeuroTreeModel` to `NeuroTree`, `LaplaceReduxModel` to `LaplaceNN`. [#444](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/444)
 - Support for `DecisionTree.jl` models and the `FeatureTweakGenerator` have been moved to an extension (`DecisionTreeExt`). [#444](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/444)
 - Updates to NeuroTreeModels extensions to incorporate breaking changes to package. [#444](https://github.com/juliatrustworthyai/CounterfactualExplanations.jl/issues/444)